able to move reducer functions to another file
