+++ docs/source/guides/adding-interactivity/components-with-state/_examples/new_example/app.py
@@ -2,30 +2,31 @@ from reactpy import component, hooks, html
 from reactpy.backend.flask import configure
 from flask import Flask
 
-# from use_counting_hook import use_counting_hook
+from use_counting_hook import use_counting_hook
 
-def reducer(count, action):
-    if action == "increment":
-        return count + 1
-    elif action == "decrement":
-        return count - 1
-    elif action == "reset":
-        return 0
-    else:
-        msg = f"Unknown action '{action}'"
-        raise ValueError(msg)
+# def reducer(count, action):
+#     if action == "increment":
+#         return count + 1
+#     elif action == "decrement":
+#         return count - 1
+#     elif action == "reset":
+#         return 0
+#     else:
+#         msg = f"Unknown action '{action}'"
+#         raise ValueError(msg)
 
 
 @component
 def Example_Component():
 
-    count, dispatch = hooks.use_reducer(reducer, 0)
+    # count, dispatch = hooks.use_reducer(reducer, 0)
+    count, dispatch = use_counting_hook(0)
 
     return html.div(
-        f"Count: {count}",
         html.button({"on_click": lambda event: dispatch("reset")}, "Reset"),
-        html.button({"on_click": lambda event: dispatch("increment")}, "+"),
-        html.button({"on_click": lambda event: dispatch("decrement")}, "-"),
+        html.button({"on_click": lambda event: dispatch("increment")}, "add"),
+        html.button({"on_click": lambda event: dispatch("decrement")}, "subtract"),
+        f"Count: {count}"
     )
